+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # What's New?
 
+## 2.1.2
+
+- Improved parsing for IAR EW toolchain messages as diagnostic problems.
+- Improved parsing for SDCC toolchain messages as diagnostic problems.
+
 ## 2.1.1
 
 - Implemented parsing for MSVC linker errors as diagnostic problems.